Use Dioxus Router for displaying views for each widget (#16)
Before, I was using a complicated setup of a shared WidgetViewState that
limited my ability to make the widget views more declarative. Now, I am
using the Dioxus Router component to display different widget views and
switch between them in a much more streamlined fashion. I was able to
remove the WidgetViewState struct and the Widget enum entirely. Now,
adding a new widget is as simple as creating the widget module with a
WIDGET_ENTRY struct and then adding it to the WIDGETS hashmap in
main.rs.