Improve layout using more Bootstrap-native spacing (#85)
I am cleaning up some of the stylesheet code to use standard Bootstrap
grid display wherever possible and only use flex display for specific
use cases such as wrapping divs and setting divs to take up remaining
space. I am not yet touching the sidebar or the homepage as those are
tinder boxes I'm not yet ready to deal with.

With this change, I'm settling on using the [Gap
spacing](https://getbootstrap.com/docs/5.3/utilities/spacing/#gap) to
create margins between elements but not at the end of them. This then
lets me apply a standard padding to the bottom of the widget view so
content doesn't get lost in the margins.