Finalized pipeline for deploying to GitHub Pages (#95)
Wow that was harder than I expected. What I've ended up settling on is
configuring yet another [HTML file](trunk.html) for the Trunk build.
I've also configured a [Trunk.toml](Trunk.toml) file with the public-url
of the Pages site and a pre-build hook that just runs the cargo build to
ensure the assets from Bootstrap and Sass are in the public folder
before running the Trunk frontend build.

I've also added caching to the build pipeline to cache Cargo assets.
This speeds up the builds significantly. As part of setting up caching,
I've uploaded the Cargo.lock file to the repo. This should help ensure
the right dependency versions are used and cached. I've also changed to
using the [upload-pages-artifact
action](https://github.com/actions/upload-pages-artifact/tree/main) for
packaging the artifact. This ensures that the artifact is in the right
format to be directly consumed by Pages when it gets deployed.