Replace chrono-tz with time-tz and improve Time Zone selection (#99)
This PR updates the DateTime converter implementation to use the
[time-tz crate](https://github.com/Yuri6037/time-tz) instead of
chrono-tz. This patches
https://github.com/esimkowitz/dev-widgets/security/dependabot/1. It also
adds functionality to get the local timezone for the target system. The
specific revision of this crate used here includes a proposed change I
have added which extends support for getting the local time zone to WASM
targets (https://github.com/Yuri6037/time-tz/pull/14).

This PR also adds support for logging, replacing use of println! in the
application logic. This way, support for logging will be equal for both
WASM and desktop targets. It also removes an unnecessary direct clap
dependency.