Update CIDR decoder to use cleaner layout, no onsubmit (#105)
I found the layout of the CIDR decoder unpleasant. It was too hard to
understand how to input the CIDR, the onsubmit form was finicky, and I
missed being able to get live feedback on the validity of the input CIDR
string. I reworked it so the details of the CIDR range are in a readonly
TextArea form and I just store the input state separately and only
update the IpCidr ref when a new valid one is present.