Update to Dioxus 0.4.0 (#112)
This PR updates all Dioxus and Dioxus-related to use the new Dioxus
0.4.0 release. The biggest change here is that the [dioxus_router
crate](https://docs.rs/dioxus-router/latest/dioxus_router/index.html)
has been completely rewritten to use a static enum-based definition
system. I really like this approach, though it forced me to rethink a
lot of how I handle widget declarations, etc. Before, I would iterate
through a PHF object of all the widget entries and create the routes in
realtime, now I have to specify them in advance and use more of the
native routing primitives to construct the sitemap. As a result of these
changes, I've removed some fields from the WidgetEntry struct and I may
end up replacing it entirely with some form of proc macro, like how the
route paths are declared in the route enums.

I've restructured the pages code so that each widget type has its own
module. This lets me give them each their own route enum, meaning that
adding new widgets can be managed at the widget type level, rather than
modifying some giant-scoped file. I don't really like the idea of
throwing all the code into a single file, even if that is one of the
advertised benefits of Dioxus.

I think the next set of changes I want to make are to rework the sidebar
and create widget-type-specific homepages. The new router layout
mechanism added a bunch of hidden `pre` tags to the sidebar that are
causing the accordion sections to display an extra lower border that I
don't like. Rather than try and address that here, I am just going to
rewrite this so it's more flexible. I want to be able to click on the
widget type and show a homepage with just widgets of that type. Clicking
on the chevron itself should be the only thing that collapses the
section. This PR adds new mod.rs files for each widget type module so I
can host this sub-home-page there and reuse the layout and logic from
the current global home page.