Fix handling of paths when refreshing from GitHub Pages (#115)
This PR overcomes a quirk of GitHub Pages where dynamic paths cannot be
processed when the page is refreshed, since they are not present in the
static site map that GitHub Pages creates. To handle this, we overload
the GitHub Pages 404 page with one that has a script to parse the
intended path from the browser and redirect to the index, applying the
dynamic path.

This also includes some cleanup of the Dioxus configs. There's a problem
with the Dioxus CLI where assets being included in the desktop bundle
use a different path structure than either cargo-bundle or building for
the Dioxus Web. There's an open issue
(https://github.com/DioxusLabs/dioxus/issues/1283) to remedy this, but
for now I am going to stick with using cargo-bundle. Bundling with the
Dioxus CLI will not package the CSS/JS.